refactor: string.h
