+++ include/nanoprintf.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define __CLANG_INTTYPES_H
+#include "nanoprintf/nanoprintf.h"
