+++ include/string.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void memzero(void* start, void* end);
+int strlen(const char* s);
+int strcmp(const char* s1, const char* s2);
+int strncmp(const char* s1, const char* s2, int n);
