+++ src/nanoprintf.c
@@ -1,0 +1,10 @@
+#include "nanoprintf.h"
+
+#define NANOPRINTF_IMPLEMENTATION
+#include "nanoprintf/nanoprintf.h"
+
+// TODO
+#include "util.h"
+void __trunctfsf2() {
+  prog_hang();
+}
